Improving the socket connection handling by using NIO.

+++ src/main/java/com/jc/multipong/bootstrap/entities/SimpleGameObject.java
@@ -5,7 +5,7 @@ package com.jc.multipong.bootstrap.entities;
  */
 public class SimpleGameObject {
 
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
 }
